feat: 어드민/아티스트 모두 하나의 AlbumList 컴포넌트로 통합
prop으로 userType을 받아 optionButton 노출 여부를 결정한다.
