+++ src/components/album/AlbumList/AlbumList.module.scss
